+++ src/app/components/qualitie.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Qualitie = ({color, name, _id}) => {
+    return (   
+        <span  key = {_id} className={"badge bg-" + color}>{name}</span>
+    )
+}
+
+export default Qualitie;
