--- src/components/users.jsx
@@ -1,74 +1,0 @@
-import React, { useState } from "react";
-import api from "../api";
-
-const Users = () => {
-  const [users, setUsers] = useState(api.users.fetchAll());
-  const handleDelete = (userID) => {
-    setUsers((users) => users.filter((user) => user._id !== userID));
-  };
-
-
-  const renderPhrase = (number) => {
-    const lastOne = Number(number.toString().slice(-1));
-    if (number > 4 && number < 15) return "человек тусанет";
-    if ([2, 3, 4].indexOf(lastOne) >= 0) return "человека тусанут";
-    if (lastOne === 1) return "человек тусанет";
-    return "человек тусанет";
-};
-
-  const renderTable = () => {
-    return (
-      <table className="table table-striped">
-        <thead>
-          <tr>
-            <th scope="col">Имя</th>
-            <th scope="col">Качества</th>
-            <th scope="col">Профессия</th>
-            <th scope="col">Встретился, раз</th>
-            <th scope="col">Оценка</th>
-            <th scope="col"></th>
-          </tr>
-        </thead>
-        <tbody>
-          {users.map((user) => (
-            <tr>
-              <td>{user.name}</td>
-              <td>
-                {user.qualities.map((qualitie) => (
-                  <span className={"badge bg-" + qualitie.color}>
-                    {qualitie.name}
-                  </span>
-                ))}
-              </td>
-              <td>{user.profession.name}</td>
-              <td>{user.completedMeetings}</td>
-              <td>{user.rate}</td>
-              <td>
-                <button
-                  className="btn btn-danger"
-                  onClick={() => handleDelete(user._id)}
-                >
-                  delete
-                </button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    );
-  };
-  return (
-    <>
-        <h2>
-            <span className={"badge " + (users.length > 0 ? "bg-primary" : "bg-danger")}>
-                    {users.length > 0
-                        ? `${users.length + " " + renderPhrase(users.length)} с тобой сегодня`
-                        : "Никто с тобой не тусанет"}
-            </span>
-        </h2>
-        {users.length>0 && renderTable()}
-    </>
-  );
-};
-
-export default Users;
