+++ src/index.js
@@ -2,13 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import Users from "./components/users";
 import "bootstrap/dist/css/bootstrap.css";
+import App from "./app/App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <Users />
+    <App />
   </React.StrictMode>
 );
